api/requirements.txt
# Core dependencies
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scipy>=1.10.0,<2.0.0
scikit-learn>=1.2.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
loguru>=0.7.0,<1.0.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0

# Web framework
fastapi>=0.100.0,<1.0.0
uvicorn>=0.20.0,<1.0.0
python-multipart>=0.0.5,<1.0.0

# Database
sqlalchemy>=2.0.0,<3.0.0
alembic>=1.11.0,<2.0.0
psycopg2-binary>=2.9.0,<3.0.0
pgvector>=0.2.0,<1.0.0
redis>=4.5.0,<5.0.0
celery>=5.3.0,<6.0.0
stripe>=7.0.0,<8.0.0

# AI/ML
torch>=2.0.0,<3.0.0
transformers>=4.30.0,<5.0.0
sentence-transformers>=2.2.0,<3.0.0
openai>=1.0.0,<2.0.0
llama-index>=0.8.0,<1.0.0
llama-index-llms-openai>=0.1.0,<1.0.0
llama-index-embeddings-openai>=0.1.0,<1.0.0
llama-index-program-openai>=0.1.0,<1.0.0
mistralai==0.4.2
dspy==2.5.43

# Audio processing
openai-whisper>=20231117  # OpenAI's Whisper for speech recognition
av>=10.0.0  # FFmpeg bindings for Python

# File processing
PyMuPDF>=1.23.0,<2.0.0
pdfminer.six>=20221105
pytesseract>=0.3.10
Pillow>=10.0.0,<11.0.0
python-pptx>=0.6.21
python-docx>=0.8.11
openpyxl>=3.1.0
python-magic>=0.4.27

# Media processing
yt-dlp>=2023.7.6
youtube-transcript-api>=0.6.1
av>=10.0.0,<11.0.0
pydub>=0.25.1
SpeechRecognition>=3.10.0
soundfile>=0.12.1
ffmpeg-python>=0.2.0

# Google services
google-api-python-client>=2.0.0,<3.0.0
google-auth>=2.0.0,<3.0.0
google-auth-httplib2>=0.1.0
google-cloud-firestore>=2.11.0,<3.0.0
google-cloud-storage>=2.10.0,<3.0.0
google-generativeai>=0.3.0,<1.0.0
firebase-admin>=6.1.0,<7.0.0

# Utilities
posthog>=3.0.0,<4.0.0
python-jose>=3.3.0,<4.0.0
requests>=2.28.0,<3.0.0
httpx>=0.24.0,<1.0.0
tqdm>=4.65.0,<5.0.0
python-dateutil>=2.8.0,<3.0.0
pytz>=2023.3
pyyaml>=6.0.0,<7.0.0
python-slugify>=8.0.0,<9.0.0

# Dependencies with specific versions due to compatibility
pyarrow>=10.0.0,<11.0.0  # Required for dspy
protobuf>=4.21.0,<5.0.0  # Required for many ML libraries
huggingface-hub>=0.16.0,<0.17.0  # Required for datasets and transformers compatibility

asyncpg==0.30.0